Detic for ONNXRuntime

+++ object_detection/detic/detic.py
@@ -340,7 +340,8 @@ def predict(net, img):
     im_h, im_w = img.shape[:2]
     img = preprocess(img)
     pred_hw = img.shape[-2:]
-    im_hw = np.array([im_h, im_w])
+    im_hw = np.array([im_h, im_w]).astype(np.int64)
+    #img[:] = 0 # test for grid sampler
 
     # feedforward
     if args.opset16:
